# frontend/public/api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Path Art Workflow API",
    "description": "Comprehensive API documentation for Path Art workflow management system. This API allows you to trigger workflows, manage executions, make decisions, update data, and interact with workflow steps programmatically.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@pathart.com"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://services.floowly.app/functions/v1",
      "description": "Floowly Services endpoint"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Company-specific API key for authentication. Obtain this from Organization Settings."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase anon key for gateway authentication. Safe to use in client-side applications."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "details": {
            "type": "string",
            "description": "Detailed error information"
          }
        },
        "required": ["error"]
      },
...
      "TranscribeAudioRequest": {
        "type": "object",
        "required": ["audio"],
        "properties": {
          "audio": {
            "type": "string",
            "format": "base64",
            "description": "Base64-encoded audio data (WebM format)"
          }
        }
      },
      "TranscribeAudioResponse": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "description": "Transcribed text from the audio"
          }
        },
        "required": ["text"]
      },
      "CreateWorkflowWithAIRequest": {
        "type": "object",
        "required": ["messages", "companyId"],
        "properties": {
          "messages": {
            "type": "array",
            "description": "Conversation messages with the AI",
            "items": {
              "type": "object",
              "properties": {
                "role": {
                  "type": "string",
                  "enum": ["user", "assistant", "system"]
                },
                "content": {
                  "type": "string"
                }
              }
            }
          },
          "companyId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "CreateWorkflowWithAIResponse": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "workflowId": {
                "type": "string",
                "format": "uuid"
              },
              "message": {
                "type": "string"
              }
            },
            "required": ["workflowId", "message"]
          },
          {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "description": "AI response message (question or information)"
              }
            },
            "required": ["message"]
          }
        ]
      }
    }
  },
  "paths": {
...
    "/transcribe-audio": {
      "post": {
        "tags": ["AI"],
        "summary": "Transcribe audio to text",
        "description": "Transcribes audio data (WebM format) to text using OpenAI Whisper. The audio should be provided as base64-encoded data.",
        "operationId": "transcribeAudio",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranscribeAudioRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Audio transcribed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscribeAudioResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing audio data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/create-workflow-with-ai": {
      "post": {
        "tags": ["AI"],
        "summary": "Create workflow with AI assistance",
        "description": "Uses AI to help create workflows through conversation. The AI will ask questions to understand your business process and then generate a complete workflow structure. Requires Bearer token authentication (user session).",
        "operationId": "createWorkflowWithAI",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkflowWithAIRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workflow created or AI response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateWorkflowWithAIResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  }
}